Move fit polygons to bounds -function into utils
