+++ src/components/MobilityPlatform/Boating/Marinas/Marinas.js
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 import MobilityPlatformContext from '../../../../context/MobilityPlatformContext';
 import { useAccessibleMap } from '../../../../redux/selectors/settings';
 import { fetchMobilityMapPolygonData } from '../../mobilityPlatformRequests/mobilityPlatformRequests';
-import { isDataValid } from '../../utils/utils';
+import { isDataValid, fitPolygonsToBounds } from '../../utils/utils';
 import MarinasContent from './components/MarinasContent';
 
 /**
@@ -40,13 +40,7 @@ const Marinas = () => {
   const renderData = isDataValid(showMarinas, marinasData);
 
   useEffect(() => {
-    if (renderData) {
-      const bounds = [];
-      marinasData.forEach((item) => {
-        bounds.push(item.geometry_coords);
-      });
-      map.fitBounds(bounds);
-    }
+    fitPolygonsToBounds(renderData, marinasData, map);
   }, [showMarinas, marinasData]);
 
   return (
